Reset timers separate from pwmio
This prevents timer leakage on builds without pwmio.

Fixes #5057
