+++ ports/atmel-samd/supervisor/port.c
@@ -77,6 +77,7 @@
 #include "samd/dma.h"
 #include "shared-bindings/microcontroller/__init__.h"
 #include "shared-bindings/rtc/__init__.h"
+#include "shared_timers.h"
 #include "reset.h"
 
 #include "supervisor/shared/safe_mode.h"
@@ -346,6 +347,9 @@ void reset_port(void) {
     #endif
     #if CIRCUITPY_PWMIO
     pwmout_reset();
+    #endif
+    #if CIRCUITPY_PWMIO || CIRCUITPY_AUDIOIO
+    reset_timers();
     #endif
 
     #if CIRCUITPY_ANALOGIO
